First pass at VolumeTableRow component
